Adding the github action that runs every weekday
The goal is to run the script every weekday. Then add CSV files and update the JSON files.
In many cases, it would fail. For example, when we can't find files on the day (It seems like the CSV files get uploaded only on business day), we would not update any of CSV/JSON files.
However, "last_updated" should always be updated, so `git add` and `git commit` wouldn't fail.